+++ src/components/display/fewlines/StandardIdentitiesSection/StandardIdentitiesSection.tsx
@@ -7,7 +7,7 @@ import {
   BoxedLink,
   NoIdentitiesBox,
 } from "../IdentitiesSection/IdentitiesSection";
-import { NeutralLink } from "../NeutralLink";
+import { NeutralLink } from "../NeutralLink/NeutralLink";
 import { Separator } from "../Separator/Separator";
 import { ShadowBox } from "../ShadowBox/ShadowBox";
 import { ShowMoreButton } from "../ShowMoreButton/ShowMoreButton";
