+++ src/components/display/fewlines/ValidateIdentityForm/ValidateIdentityForm.tsx
@@ -7,7 +7,7 @@ import { Box } from "../Box/Box";
 import { Button, ButtonVariant } from "../Button/Button";
 import { Form } from "../Form/Form";
 import { Input } from "../Input/Input";
-import { NeutralLink } from "../NeutralLink";
+import { NeutralLink } from "../NeutralLink/NeutralLink";
 import { IdentityTypes } from "@lib/@types/Identity";
 import {
   InvalidValidationCode,
